Creation of classes and implementation of starting set for the class

+++ Source/GS_RPGGame/Public/Components/GSLevelComponent.h
@@ -37,7 +37,7 @@ public:
 	* Adds levels
 	*/
 	UFUNCTION(BlueprintCallable, Category = "Level")
-	void AddLevels(float Levels);
+	void AddLevels(int Levels);
 
 	/**
 	* Returns wether the current level is the max
